tests/integration/requirements.txt
aiohttp == 3.11.7
cosl == 0.1.0
git+https://github.com/canonical/saml-test-idp.git
jubilant == 1.6.2
juju == 3.6.1.2
minio == 7.2.14
nest_asyncio == 1.6.0
ops >= 1.5.0
pytest-operator >= 0.32.0
tenacity == 9.0.0
playwright == 1.53.0
